Specify conditions to return true or false
Specify true or false conditions for the checkValue
function so that if the elements in removeArray and
editArray match, the return is false.

Conditional statement in the nested for loop will
then specify that a false return should call the
filterArray function & filter out the element to be
removed from the original array, returning the newly
edited array.