Remove nested loop and checkRemoves function
Removed nested loop b/c checkRemoves function is redundant with
functional removeItems function as they are both trying to find the
matching values and deleting them from the array.

Filter function now works as expected, but tests with parameters that
are not present in the original array are returning incorrect arrays
including undefined and strings.